requirements.txt
Django>=4.2
djangorestframework>=3.14
Pillow>=10.0
# mysqlclient>=2.1.0  # Uncomment for MySQL support
